Added corner collision handling. This fixes wall humping significantly, except now it can happen only through single walls and not corners.

+++ FOP.cc
@@ -134,7 +134,7 @@ FOP* FOP::Character(int x, int y, Tile* ref)
 
 void FOP::startAnimation()
 {
-	if(oldTime + frameRate > SDL_GetTicks())
+	if((unsigned)oldTime + (unsigned)frameRate > SDL_GetTicks())
 	{
 		return;
 	}
